Move service worker implementation to comlink worker
