+++ packages/pyodide-kernel/src/worker.ts
@@ -5,13 +5,11 @@ import type Pyodide from 'pyodide';
 
 import type { DriveFS } from '@jupyterlite/contents';
 
-import { URLExt } from '@jupyterlab/coreutils';
-
 import { KernelMessage } from '@jupyterlab/services';
 
 import type { IPyodideWorkerKernel } from './tokens';
 
-export class PyodideRemoteKernel {
+export abstract class PyodideRemoteKernel {
   constructor() {
     this._initialized = new Promise((resolve, reject) => {
       this._initializer = { resolve, reject };
@@ -455,59 +453,16 @@ export class PyodideRemoteKernel {
   }
 
   /**
-   * Send a input request to the front-end via the service worker and block until the
-   * reply is received.
+   * Send a input request to the front-end and block until the reply is received.
    *
    * @param prompt the text to show at the prompt
    * @param password Is the request for a password?
+   * @returns String value from the input reply message, or undefined if there is none.
    */
-  sendInputRequest(prompt: string, password: boolean): string | undefined {
-    const parentHeader = this.formatResult(this._kernel._parent_header)['header'];
-
-    // Filling out the input_request message fields based on jupyterlite BaseKernet.inputRequest
-    const inputRequest = KernelMessage.createMessage<KernelMessage.IInputRequestMsg>({
-      channel: 'stdin',
-      msgType: 'input_request',
-      session: parentHeader?.session ?? '',
-      parentHeader: parentHeader,
-      content: {
-        prompt,
-        password,
-      },
-    });
-
-    try {
-      if (!this._options) {
-        throw new Error('Kernel options not set');
-      }
-
-      const { baseUrl, browsingContextId } = this._options;
-      if (!browsingContextId) {
-        throw new Error('Kernel browsingContextId not set');
-      }
-
-      const xhr = new XMLHttpRequest();
-      const url = URLExt.join(baseUrl, '/stdin/kernel');
-      xhr.open('POST', url, false); // Synchronous XMLHttpRequest
-      const msg = JSON.stringify({
-        browsingContextId,
-        data: inputRequest,
-      });
-      // Send input request, this blocks until the input reply is received.
-      xhr.send(msg);
-      const inputReply = JSON.parse(xhr.response as string);
-
-      if ('error' in inputReply) {
-        // Service worker may return an error instead of an input reply message.
-        throw new Error(inputReply['error']);
-      }
-
-      return inputReply.content?.value;
-    } catch (err) {
-      console.warn(`Failed to request stdin via service worker: ${err}`);
-      return undefined;
-    }
-  }
+  protected abstract sendInputRequest(
+    prompt: string,
+    password: boolean,
+  ): string | undefined;
 
   getpass(prompt: string): string | undefined {
     prompt = typeof prompt === 'undefined' ? '' : prompt;
